add support for arbitrary maps, and an example using them.
this will be extended until it replaces the current PageData struct -- I don't want a single hint of html idiosyncrasies like a "title" field leaking into the core logic of the templating.  the Path type and its helpers might survive; paths are both special enough and common enough that it should be non-iritating to handle them in templates without importing the whole set of functions from the filepath library.