change the flatsite Map type to a struct instead of just a typedef over normal maps.  new example using "{{with NewMap}}" in the template.
turns out the golang template library saw the old flatsite Map typedef as still being a map... and saw it as being "empty" when you made a new one.  Which meant you couldn't use a "with" block on a new Map.  I want to use a "with" block on a new Map.  this wrapper struct makes that legal.