return Map to being a typedef instead of a struct, in order to get easy traverses again.
interestingly, the template system still checks for methods before checking if something is a map and thus possibly a key... so methods on our Map type can indeed be called as methods, while also giving us easy key traverses.  in other words, the .Set operation can still easily be performed at the same time as we gain easy traverses.  you can even set a "Set" key... you just won't be able to get it back out with the easy traverse syntax (but you will by calling .Get "Set"), which makes perfect sense.

the downside is this necessarily loses the ability to use a "{{with NewMap}}" block.  I can't come up with any way to trick the processing of the with clause into making an exception without actually putting junk values in the map.  The with clause processing uses a the template package's concept of isTrue, which in turn punts to reflect's Len for map kinds, which in turn punts to maplen, which goes straight to native code, do not pass go, do not collect 200 dollars.