regain the "{{with NewMap}}" ability with a hack.
injects an empty key into a NewMap.  remove it again as soon as anything is added.  this is somewhat odd behavior, but I also can't imagine why you'd ever allocate a map without putting a single thing in it, and at a practical level I find it even more odd to be unable to declare a new block with a new map, so the trade seems valid.