alter the top-level Set function to return the value set.
this makes it possible to do stuff like `{{ with (Set . "Submap" NewMap) }}`, which is a really elegant way of created a new sub-map in an existing context, and switching to adding content to the new submap in the same breath.  (the alternative is to save the dot reference as a named variable, then make the new map, then do a separate set step.  clearly, this much is shorter.)

the .Set function on Map still returns a blank string, since it's typically used where output would be undesirable and piping the .Set call to `printf ""` would be ridiculous.